+++ src/main/kotlin/com/github/chenlijun99/jetbrainsassignment/Constants.kt
@@ -1,0 +1,5 @@
+package com.github.chenlijun99.jetbrainsassignment
+
+internal object Constants {
+    const val NOTIFICATION_GROUP = "com.github.chenlijun99.jetbrainsassignment.NotificationGroup"
+}