+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 ## [Unreleased]
 
+### Fixed
+
+- i18n: Removed useless "EN" suffix from english translations
+
 ## [0.0.1] - 2025-10-19
 
 ### Added