+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ## [Unreleased]
 
+## [0.1.0] - 2025-10-20
+
 ### Added
 
 - Introduced more informative message about compression result: original file size, compressed file size and compression percentage.
@@ -23,5 +25,7 @@
 - First MVP
 - Added support for internationalization
 
-[Unreleased]: https://github.com/chenlijun99/jetbrains-assignment/compare/v0.0.1...HEAD
+[Unreleased]: https://github.com/chenlijun99/jetbrains-assignment/compare/v0.1.0...HEAD
+[0.1.0]: https://github.com/chenlijun99/jetbrains-assignment/compare/v0.0.1...v0.1.0
+[0.0.2]: https://github.com/chenlijun99/jetbrains-assignment/compare/v0.0.1...v0.0.2
 [0.0.1]: https://github.com/chenlijun99/jetbrains-assignment/commits/v0.0.1